# aifo/slidescope/internal/server/public/openapi.yaml
basePath: /api/v1
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  handlers.LoginInput:
    properties:
      password:
        example: password123
        type: string
      username:
        example: testuser
        type: string
    required:
      - password
      - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A digital pathology slide management and viewing API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SlideScope API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate user credentials and return JWT tokens in cookies
      parameters:
        - description: User credentials
          in: body
          name: credentials
          required: true
          schema:
            $ref: '#/definitions/handlers.LoginInput'
      produces:
        - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: User login
      tags:
        - authentication
  /api/v1/auth/logout:
    post:
      description: Clear authentication cookies and log out the user
      produces:
        - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: User logout
      tags:
        - authentication
  /api/v1/auth/me:
    get:
      description: Retrieve information about the currently authenticated user
      produces:
        - application/json
      responses:
        "200":
          description: Current user information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get current user information
      tags:
        - authentication
  /api/v1/auth/refresh:
    post:
      description: Use refresh token to generate new access and refresh tokens
      produces:
        - application/json
      responses:
        "200":
          description: Token refresh successful
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh authentication tokens
      tags:
        - authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"